pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Tom Downey <tom.downey@sky.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.0"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.11.0"
gpt-index = "^0.8.36"
pypdf = "^3.16.2"
google-cloud-aiplatform = "^1.35.0"
google-cloud-texttospeech = "^2.14.2"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
